Split ust/tst files into expressions and declarations
