+++ lang/syntax/src/trees/generic/lookup_table.rs
@@ -1,6 +1,6 @@
 use std::fmt;
 
-use super::def::*;
+use super::exp::*;
 use crate::common::*;
 
 #[derive(Clone, Copy, Debug)]
