+++ lang/syntax/src/trees/generic/mod.rs
@@ -1,6 +1,8 @@
-mod def;
+mod decls;
+mod exp;
 mod lookup;
 pub mod lookup_table;
 
-pub use def::*;
+pub use decls::*;
+pub use exp::*;
 pub use lookup::*;
